Shift wait on full from File in corresponding classes.
This cleans up things. However the wait abort has to happen now
in the classes. So I changed the close function to not wait till
size=0, but to just stop further read/writes. This can lead to problems
at close of write files. Need to write tests for this later.

Signed-off-by: Tobias Lorenz <tobias.lorenz@gmx.net>